Prevent message loops on FDC3 channels (#3)
Watson has been having fun with message loops on FDC3 channels. He's comparing the data in messages received with one's just sent which seems fragile. We could instead add a source field to each message and check it before calling that handler (so no client will process a message it sent). 

The code in this PR is not tested, but provided to illustrate the idea. 
Felt the need to do a little deduplication at the same time, which I think makes it easier to read.